+++ lib/src/dom_tools_base.dart
@@ -4,6 +4,72 @@ import 'dart:math';
 
 import 'package:swiss_knife/swiss_knife.dart';
 
+/// Gets the [element] value depending of identified type.
+///
+/// If the resolved value is null or empty, and def is not null,
+/// it will return [def].
+String getElementValue(Element element, [String def]) {
+  if (element == null) return def ;
+
+  String value ;
+
+  if (element is InputElement) {
+    value = element.value ;
+  }
+  else if (element is CanvasImageSource) {
+    value = getElementSRC(element) ;
+  }
+  else if (element is CheckboxInputElement) {
+    value = element.checked ? 'true' : 'false' ;
+  }
+  else if (element is TextAreaElement) {
+    value = element.value ;
+  }
+  else if ( isElementWithSRC(element) ) {
+    value = getElementSRC(element) ;
+  }
+  else if ( isElementWithHREF(element) ) {
+    value = getElementHREF(element) ;
+  }
+  else {
+    value = element.text ;
+  }
+
+  return def != null && isEmptyObject(value) ? def : value ;
+}
+
+/// Sets the [element] [value] depending of the identified type.
+bool setElementValue(Element element, String value) {
+  if (element == null) return false ;
+
+  if (element is InputElement) {
+    element.value = value ;
+    return true ;
+  }
+  else if (element is CanvasImageSource) {
+    return setElementSRC(element, value) ;
+  }
+  else if (element is CheckboxInputElement) {
+    element.checked = parseBool(value) ;
+    return true ;
+  }
+  else if (element is TextAreaElement) {
+    element.value = value ;
+    return true ;
+  }
+  else if ( isElementWithSRC(element) ) {
+    return setElementSRC(element, value) ;
+  }
+  else if ( isElementWithHREF(element) ) {
+    return setElementHREF(element, value) ;
+  }
+  else {
+    element.text = value ;
+    return true ;
+  }
+
+}
+
 /// Returns a value from an [Element].
 typedef ElementValueGetter<T> = T Function(Element element);
 
@@ -36,6 +102,38 @@ String getElementHREF(Element element) {
   return null;
 }
 
+/// Sets [element] [href] depending of the identified type.
+bool setElementHREF(Element element, String href) {
+  if (element is LinkElement) {
+    element.href = href ;
+    return true ;
+  }
+  else if (element is AnchorElement) {
+    element.href = href ;
+    return true ;
+  }
+  else if (element is BaseElement) {
+    element.href = href ;
+    return true ;
+  }
+  else if (element is AreaElement) {
+    element.href = href ;
+    return true ;
+  }
+
+  return false;
+}
+
+/// Returns [true] if [element] type can have `href` attribute.
+bool isElementWithHREF(Element element) {
+  if (element is LinkElement) return true;
+  if (element is AnchorElement) return true;
+  if (element is BaseElement) return true;
+  if (element is AreaElement) return true;
+
+  return false;
+}
+
 /// Returns `src` value for different [Element] types.
 String getElementSRC(Element element) {
   if (element is ImageElement) return element.src;
@@ -52,6 +150,70 @@ String getElementSRC(Element element) {
   if (element is ImageButtonInputElement) return element.src;
 
   return null;
+}
+
+
+/// Sets the [element] [src] depending of the identified type.
+bool setElementSRC(Element element, String src) {
+  if (element == null) return false ;
+
+  if (element is ImageElement) {
+    element.src = src ;
+    return true ;
+  }
+  else if (element is ScriptElement) {
+    element.src = src ;
+    return true ;
+  }
+  else if (element is InputElement) {
+    element.src = src ;
+    return true ;
+  }
+  else if (element is MediaElement) {
+    element.src = src ;
+    return true ;
+  }
+  else if (element is EmbedElement) {
+    element.src = src ;
+    return true ;
+  }
+  else if (element is IFrameElement) {
+    element.src = src ;
+    return true ;
+  }
+  else if (element is SourceElement) {
+    element.src = src ;
+    return true ;
+  }
+  else if (element is TrackElement) {
+    element.src = src ;
+    return true ;
+  }
+  else if (element is ImageButtonInputElement) {
+    element.src = src ;
+    return true ;
+  }
+  else {
+    return false ;
+  }
+}
+
+/// Returns [true] if [element] type can have `src` attribute.
+bool isElementWithSRC(Element element) {
+  if (element is ImageElement) return true ;
+  if (element is ScriptElement) return true ;
+  if (element is InputElement) return true ;
+
+  if (element is MediaElement) return true ;
+  if (element is EmbedElement) return true ;
+
+  if (element is IFrameElement) return true ;
+  if (element is SourceElement) return true ;
+  if (element is TrackElement) return true ;
+
+  if (element is ImageButtonInputElement) return true ;
+
+  return false;
 }
 
 /// Selects an [Element] in DOM with [tag] and [href].