+++ lib/src/dom_tools_document.dart
@@ -239,3 +239,22 @@ String markdownToHtml(String markdown,
 
   return markdownHtml;
 }
+
+/// Downloads [content] of type [mimeType], saving a file with [fileName].
+void downloadContent(List<String> content, MimeType mimeType, String fileName) {
+  var blob = Blob(content, mimeType.toString()) ;
+
+  var fileLink = AnchorElement() ;
+  fileLink.style.display = 'none' ;
+
+  if ( isNotEmptyObject(fileName) ) fileLink.download = fileName ;
+  fileLink.href = Url.createObjectUrlFromBlob(blob) ;
+
+  fileLink.onClick.listen((event) {
+    fileLink.remove();
+  });
+
+  document.body.append( fileLink ) ;
+
+  fileLink.click() ;
+}